backend/requirements.txt
django-cors-headers==4.4.0
Django==5.1.1
djangorestframework==3.15.2
gunicorn==23.0.0
numpy==2.1.2
pandas==2.2.3
psycopg[binary]==3.2.10
PyJWT==2.9.0
python-dotenv==1.0.1
reportlab==4.2.5
scikit-learn==1.5.2
supabase==2.5.0
whitenoise==6.6.0
